Move Cache class into caching package
