+++ jollyday-core/src/main/java/de/focus_shift/HolidayManager.java
@@ -4,8 +4,8 @@ import de.focus_shift.caching.HolidayManagerValueHandler;
 import de.focus_shift.configuration.ConfigurationProviderManager;
 import de.focus_shift.parser.functions.CalendarToLocalDate;
 import de.focus_shift.spi.ConfigurationService;
-import de.focus_shift.util.Cache;
-import de.focus_shift.util.Cache.ValueHandler;
+import de.focus_shift.caching.Cache;
+import de.focus_shift.caching.Cache.ValueHandler;
 import de.focus_shift.util.CalendarUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
